Added Task5

+++ Tyuiu.KozachokNA.Sprint0.Task4.V0.Lib/DataService.cs
@@ -10,7 +10,7 @@ namespace Tyuiu.KozachokNA.Sprint0.Task4.V0.Lib
         { return a - b; }
         public static int Multiplication(int a, int b)
         { return a * b; }
-        public static int Division(int a, int b)
+        public static double Division(int a, int b)
         { return a / b; }
     }
 }
